web/api/requirements.txt
python-multipart
pyjwt[crypto]==2.10.*
fastapi==0.116.0
uvicorn==0.34.2
loguru==0.7.3
sqlmodel==0.0.24
pydantic==2.*
